# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Channel:
    properties:
      config:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      qr_code:
        type: string
      qr_expires_at:
        type: string
      session:
        description: session identifier for WhatsApp integration
        type: string
      status:
        description: disconnected, connecting, connected
        type: string
      tenant_id:
        type: string
      type:
        description: whatsapp, webchat, etc.
        type: string
      updated_at:
        type: string
      webhook_token:
        type: string
      webhook_url:
        type: string
    required:
    - name
    - session
    - type
    type: object
  models.ChannelListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.SwaggerChannelWithConversationCount'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.SwaggerChannelWithConversationCount:
    properties:
      access_token:
        type: string
      business_phone:
        type: string
      conversation_count:
        type: integer
      created_at:
        type: string
      id:
        type: string
      instance_id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
      tenant_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      webhook_token:
        type: string
      webhook_url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Multi-tenant SaaS para vendas e atendimento via WhatsApp
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: IAFarma API
  version: "1.0"
paths:
...
  /admin/export/tenants:
    get:
      description: Get a list of all tenants to choose from for product export
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all tenants for export selection
      tags:
      - admin-export
  /admin/export/tenants/{tenant_id}/products:
    get:
      description: Export all products from a specific tenant to CSV format for backup
        purposes
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Export tenant products to CSV
      tags:
      - admin-export
...
  /admin/tenants/{tenant_id}/channels:
    get:
      consumes:
      - application/json
      description: List all channels for a specific tenant - only accessible by system
        admin
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChannelListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List channels for a specific tenant (Admin only)
      tags:
      - admin
      - channels
    post:
      consumes:
      - application/json
      description: Create a new channel for a specific tenant - only accessible by
        system admin
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - description: Channel data
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/models.Channel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Channel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a channel for a specific tenant (Admin only)
      tags:
      - admin
      - channels
  /admin/tenants/{tenant_id}/channels/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a channel for a specific tenant - only accessible by system
        admin
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a channel for a specific tenant (Admin only)
      tags:
      - admin
      - channels
    put:
      consumes:
      - application/json
      description: Update a channel for a specific tenant - only accessible by system
        admin
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel data
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/models.Channel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a channel for a specific tenant (Admin only)
      tags:
      - admin
      - channels
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"